pyproject.toml
[tool.poetry]
name = "realyoloflaskdemodevelope"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.0.1"
torch = "^2.0.1"
pillow = "^9.5.0"
opencv-python = "^4.7.0.72"
psutil = "^5.9.5"
torchvision = "^0.15.2"
pyyaml = "^6.0"
tqdm = "^4.65.0"
ultralytics = "^8.0.105"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
